--- feature-forms/src/main/java/com/esri/arcgismaps/sample/featureforms/components/bottomsheet/BottomSheetScaffold.kt
@@ -1,454 +1,0 @@
-/*
- * Copyright 2023 The Android Open Source Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- *
- * Modifications copyright (C) 2023 Esri Inc
- */
-
-package com.esri.arcgismaps.sample.featureforms.components.bottomsheet
-
-import androidx.compose.foundation.gestures.Orientation
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.ColumnScope
-import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.requiredHeightIn
-import androidx.compose.foundation.layout.requiredWidth
-import androidx.compose.material3.BottomSheetDefaults
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.SnackbarHost
-import androidx.compose.material3.SnackbarHostState
-import androidx.compose.material3.Surface
-import androidx.compose.material3.contentColorFor
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.Stable
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.rememberCoroutineScope
-import androidx.compose.ui.Alignment.Companion.CenterHorizontally
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.Shape
-import androidx.compose.ui.input.nestedscroll.nestedScroll
-import androidx.compose.ui.layout.SubcomposeLayout
-import androidx.compose.ui.platform.LocalDensity
-import androidx.compose.ui.semantics.collapse
-import androidx.compose.ui.semantics.dismiss
-import androidx.compose.ui.semantics.expand
-import androidx.compose.ui.semantics.semantics
-import androidx.compose.ui.unit.Dp
-import kotlinx.coroutines.launch
-import java.lang.Float.max
-import kotlin.math.roundToInt
-
-/**
- * <a href="https://m3.material.io/components/bottom-sheets/overview" class="external" target="_blank">Material Design standard bottom sheet scaffold</a>.
- *
- * Standard bottom sheets co-exist with the screen’s main UI region and allow for simultaneously
- * viewing and interacting with both regions. They are commonly used to keep a feature or
- * secondary content visible on screen when content in main UI region is frequently scrolled or
- * panned.
- *
- * ![Bottom sheet image](https://developer.android.com/images/reference/androidx/compose/material3/bottom_sheet.png)
- *
- * This component provides API to put together several material components to construct your
- * screen, by ensuring proper layout strategy for them and collecting necessary data so these
- * components will work together correctly.
- *
- * A simple example of a standard bottom sheet looks like this:
- *
- * @sample androidx.compose.material3.samples.SimpleBottomSheetScaffoldSample
- *
- * @param sheetContent the content of the bottom sheet
- * @param modifier the [Modifier] to be applied to this scaffold
- * @param scaffoldState the state of the bottom sheet scaffold
- * @param sheetPeekHeight the height of the bottom sheet when it is collapsed
- * @param sheetShape the shape of the bottom sheet
- * @param sheetExpansionHeight the height for the bottom sheet when it is partially expanded and
- * expanded
- * @param sheetContainerColor the background color of the bottom sheet
- * @param sheetContentColor the preferred content color provided by the bottom sheet to its
- * children. Defaults to the matching content color for [sheetContainerColor], or if that is
- * not a color from the theme, this will keep the same content color set above the bottom sheet.
- * @param sheetTonalElevation the tonal elevation of the bottom sheet
- * @param sheetShadowElevation the shadow elevation of the bottom sheet
- * @param sheetDragHandle optional visual marker to pull the scaffold's bottom sheet
- * @param sheetSwipeEnabled whether the sheet swiping is enabled and should react to the user's
- * input
- * @param topBar top app bar of the screen, typically a [SmallTopAppBar]
- * @param snackbarHost component to host [Snackbar]s that are pushed to be shown via
- * [SnackbarHostState.showSnackbar], typically a [SnackbarHost]
- * @param containerColor the color used for the background of this scaffold. Use [Color.Transparent]
- * to have no color.
- * @param contentColor the preferred color for content inside this scaffold. Defaults to either the
- * matching content color for [containerColor], or to the current [LocalContentColor] if
- * [containerColor] is not a color from the theme.
- * @param content content of the screen. The lambda receives a [PaddingValues] that should be
- * applied to the content root via [Modifier.padding] and [Modifier.consumeWindowInsets] to
- * properly offset top and bottom bars. If using [Modifier.verticalScroll], apply this modifier to
- * the child of the scroll, and not on the scroll itself.
- */
-@Composable
-@ExperimentalMaterial3Api
-fun BottomSheetScaffold(
-    sheetContent: @Composable ColumnScope.() -> Unit,
-    modifier: Modifier = Modifier,
-    scaffoldState: BottomSheetScaffoldState = rememberBottomSheetScaffoldState(),
-    sheetPeekHeight: Dp = BottomSheetDefaults.SheetPeekHeight,
-    sheetShape: Shape = BottomSheetDefaults.ExpandedShape,
-    sheetExpansionHeight: SheetExpansionHeight = SheetExpansionHeight(),
-    sheetContainerColor: Color = BottomSheetDefaults.ContainerColor,
-    sheetContentColor: Color = contentColorFor(sheetContainerColor),
-    sheetTonalElevation: Dp = BottomSheetDefaults.Elevation,
-    sheetShadowElevation: Dp = BottomSheetDefaults.Elevation,
-    sheetDragHandle: @Composable (() -> Unit)? = { BottomSheetDefaults.DragHandle() },
-    sheetSwipeEnabled: Boolean = true,
-    topBar: @Composable (() -> Unit)? = null,
-    snackbarHost: @Composable (SnackbarHostState) -> Unit = { SnackbarHost(it) },
-    containerColor: Color = MaterialTheme.colorScheme.surface,
-    contentColor: Color = contentColorFor(containerColor),
-    content: @Composable (PaddingValues) -> Unit
-) {
-    BottomSheetScaffoldLayout(
-        modifier = modifier,
-        topBar = topBar,
-        body = content,
-        snackbarHost = {
-            snackbarHost(scaffoldState.snackbarHostState)
-        },
-        sheetPeekHeight = sheetPeekHeight,
-        sheetOffset = { scaffoldState.bottomSheetState.requireOffset() },
-        sheetState = scaffoldState.bottomSheetState,
-        containerColor = containerColor,
-        contentColor = contentColor,
-        bottomSheet = { layoutHeight ->
-            StandardBottomSheet(
-                state = scaffoldState.bottomSheetState,
-                peekHeight = sheetPeekHeight,
-                expansionHeight = sheetExpansionHeight,
-                sheetSwipeEnabled = sheetSwipeEnabled,
-                layoutHeight = layoutHeight.toFloat(),
-                sheetWidth = BottomSheetMaxWidth,
-                shape = sheetShape,
-                containerColor = sheetContainerColor,
-                contentColor = sheetContentColor,
-                tonalElevation = sheetTonalElevation,
-                shadowElevation = sheetShadowElevation,
-                dragHandle = sheetDragHandle,
-                content = sheetContent
-            )
-        }
-    )
-}
-
-/**
- * State of the [BottomSheetScaffold] composable.
- *
- * @param bottomSheetState the state of the persistent bottom sheet
- * @param snackbarHostState the [SnackbarHostState] used to show snackbars inside the scaffold
- */
-@ExperimentalMaterial3Api
-@Stable
-class BottomSheetScaffoldState(
-    val bottomSheetState: SheetState,
-    val snackbarHostState: SnackbarHostState
-)
-
-/**
- * Create and [remember] a [BottomSheetScaffoldState].
- *
- * @param bottomSheetState the state of the standard bottom sheet. See
- * [rememberStandardBottomSheetState]
- * @param snackbarHostState the [SnackbarHostState] used to show snackbars inside the scaffold
- */
-@Composable
-@ExperimentalMaterial3Api
-fun rememberBottomSheetScaffoldState(
-    bottomSheetState: SheetState = rememberStandardBottomSheetState(),
-    snackbarHostState: SnackbarHostState = remember { SnackbarHostState() }
-): BottomSheetScaffoldState {
-    return remember(bottomSheetState, snackbarHostState) {
-        BottomSheetScaffoldState(
-            bottomSheetState = bottomSheetState,
-            snackbarHostState = snackbarHostState
-        )
-    }
-}
-
-/**
- * Create and [remember] a [SheetState] for [BottomSheetScaffold].
- *
- * @param initialValue the initial value of the state. Should be either [PartiallyExpanded] or
- * [Expanded] if [skipHiddenState] is true
- * @param confirmValueChange optional callback invoked to confirm or veto a pending state change
- * @param [skipHiddenState] whether Hidden state is skipped for [BottomSheetScaffold]
- */
-@Composable
-@ExperimentalMaterial3Api
-fun rememberStandardBottomSheetState(
-    initialValue: SheetValue = SheetValue.PartiallyExpanded,
-    confirmValueChange: (SheetValue) -> Boolean = { true },
-    skipHiddenState: Boolean = true,
-) = rememberSheetState(false, confirmValueChange, initialValue, skipHiddenState)
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun StandardBottomSheet(
-    state: SheetState,
-    peekHeight: Dp,
-    expansionHeight: SheetExpansionHeight,
-    layoutHeight: Float,
-    sheetWidth: Dp,
-    sheetSwipeEnabled: Boolean = true,
-    shape: Shape = BottomSheetDefaults.ExpandedShape,
-    containerColor: Color = BottomSheetDefaults.ContainerColor,
-    contentColor: Color = contentColorFor(containerColor),
-    tonalElevation: Dp = BottomSheetDefaults.Elevation,
-    shadowElevation: Dp = BottomSheetDefaults.Elevation,
-    dragHandle: @Composable (() -> Unit)? = { BottomSheetDefaults.DragHandle() },
-    content: @Composable ColumnScope.() -> Unit
-) {
-    val scope = rememberCoroutineScope()
-    val peekHeightPx = with(LocalDensity.current) { peekHeight.toPx() }
-    val orientation = Orientation.Vertical
-    // Callback that is invoked when the anchors have changed.
-    val anchorChangeHandler = remember(state, scope) {
-        BottomSheetScaffoldAnchorChangeHandler(
-            state = state,
-            animateTo = { target, velocity ->
-                scope.launch {
-                    state.swipeableState.animateTo(
-                        target, velocity = velocity
-                    )
-                }
-            },
-            snapTo = { target ->
-                scope.launch { state.swipeableState.snapTo(target) }
-            }
-        )
-    }
-    Surface(
-        modifier = Modifier
-            .requiredWidth(sheetWidth)
-            .fillMaxWidth()
-            .requiredHeightIn(min = peekHeight)
-            .nestedScroll(
-                remember(state.swipeableState) {
-                    ConsumeSwipeWithinBottomSheetBoundsNestedScrollConnection(
-                        sheetState = state,
-                        orientation = orientation,
-                        onFling = { scope.launch { state.settle(it) } }
-                    )
-                }
-            )
-            .swipeableV2(
-                state = state.swipeableState,
-                orientation = orientation,
-                enabled = sheetSwipeEnabled
-            )
-            .swipeAnchors(
-                state.swipeableState,
-                possibleValues = setOf(
-                    SheetValue.Hidden,
-                    SheetValue.Minimized,
-                    SheetValue.PartiallyExpanded,
-                    SheetValue.Expanded
-                ),
-                anchorChangeHandler = anchorChangeHandler
-            ) { value, sheetSize ->
-                when (value) {
-                    SheetValue.PartiallyExpanded -> (layoutHeight - peekHeightPx - (sheetSize.height * expansionHeight.partialHeightFraction))
-                    SheetValue.Expanded -> if (sheetSize.height == peekHeightPx.roundToInt()) {
-                        null
-                    } else {
-                        max(
-                            0f,
-                            (layoutHeight - (sheetSize.height * expansionHeight.fullHeightFraction))
-                        )
-                    }
-
-                    SheetValue.Minimized -> layoutHeight - peekHeightPx
-                    SheetValue.Hidden -> layoutHeight
-                }
-            },
-        shape = shape,
-        color = containerColor,
-        contentColor = contentColor,
-        tonalElevation = tonalElevation,
-        shadowElevation = shadowElevation,
-    ) {
-        Column(Modifier.fillMaxWidth()) {
-            if (dragHandle != null) {
-                val partialExpandActionLabel =
-                    getString(Strings.BottomSheetPartialExpandDescription)
-                val dismissActionLabel = getString(Strings.BottomSheetDismissDescription)
-                val expandActionLabel = getString(Strings.BottomSheetExpandDescription)
-                Box(
-                    Modifier
-                        .align(CenterHorizontally)
-                        .semantics(mergeDescendants = true) {
-                            with(state) {
-                                // Provides semantics to interact with the bottomsheet if there is more
-                                // than one anchor to swipe to and swiping is enabled.
-                                if (swipeableState.anchors.size > 1 && sheetSwipeEnabled) {
-                                    if (currentValue == SheetValue.PartiallyExpanded) {
-                                        if (swipeableState.confirmValueChange(SheetValue.Expanded)) {
-                                            expand(expandActionLabel) {
-                                                scope.launch { expand() }; true
-                                            }
-                                        }
-                                    } else {
-                                        if (swipeableState.confirmValueChange(SheetValue.PartiallyExpanded)) {
-                                            collapse(partialExpandActionLabel) {
-                                                scope.launch { partialExpand() }; true
-                                            }
-                                        }
-                                    }
-                                    if (!state.skipHiddenState) {
-                                        dismiss(dismissActionLabel) {
-                                            scope.launch { hide() }
-                                            true
-                                        }
-                                    }
-                                }
-                            }
-                        },
-                ) {
-                    dragHandle()
-                }
-            }
-            content()
-        }
-    }
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-private fun BottomSheetScaffoldLayout(
-    modifier: Modifier,
-    topBar: @Composable (() -> Unit)?,
-    body: @Composable (innerPadding: PaddingValues) -> Unit,
-    bottomSheet: @Composable (layoutHeight: Int) -> Unit,
-    snackbarHost: @Composable () -> Unit,
-    sheetPeekHeight: Dp,
-    sheetOffset: () -> Float,
-    sheetState: SheetState,
-    containerColor: Color,
-    contentColor: Color,
-) {
-    SubcomposeLayout { constraints ->
-        val layoutWidth = constraints.maxWidth
-        val layoutHeight = constraints.maxHeight
-        val looseConstraints = constraints.copy(minWidth = 0, minHeight = 0)
-
-        val topBarPlaceable = topBar?.let {
-            subcompose(BottomSheetScaffoldLayoutSlot.TopBar) { topBar() }[0]
-                .measure(looseConstraints)
-        }
-        val topBarHeight = topBarPlaceable?.height ?: 0
-
-        val sheetPlaceable = subcompose(BottomSheetScaffoldLayoutSlot.Sheet) {
-            bottomSheet(layoutHeight - topBarHeight)
-        }[0].measure(looseConstraints.copy(maxHeight = layoutHeight - topBarHeight))
-        val sheetOffsetY = sheetOffset().roundToInt() + topBarHeight
-        val sheetOffsetX = Integer.max(0, (layoutWidth - sheetPlaceable.width) / 2)
-
-        val bodyConstraints = looseConstraints.copy(maxHeight = layoutHeight - topBarHeight)
-        val bodyPlaceable = subcompose(BottomSheetScaffoldLayoutSlot.Body) {
-            Surface(
-                modifier = modifier,
-                color = containerColor,
-                contentColor = contentColor,
-            ) { body(PaddingValues(bottom = sheetPeekHeight)) }
-        }[0].measure(bodyConstraints)
-
-        val snackbarPlaceable = subcompose(BottomSheetScaffoldLayoutSlot.Snackbar, snackbarHost)[0]
-            .measure(looseConstraints)
-        val snackbarOffsetX = (layoutWidth - snackbarPlaceable.width) / 2
-        val snackbarOffsetY = when (sheetState.currentValue) {
-            SheetValue.PartiallyExpanded -> sheetOffsetY - snackbarPlaceable.height
-            SheetValue.Expanded, SheetValue.Hidden, SheetValue.Minimized -> layoutHeight - snackbarPlaceable.height
-        }
-
-        layout(layoutWidth, layoutHeight) {
-            // Placement order is important for elevation
-            bodyPlaceable.placeRelative(0, topBarHeight)
-            topBarPlaceable?.placeRelative(0, 0)
-            sheetPlaceable.placeRelative(sheetOffsetX, sheetOffsetY)
-            snackbarPlaceable.placeRelative(snackbarOffsetX, snackbarOffsetY)
-        }
-    }
-}
-
-@ExperimentalMaterial3Api
-private fun BottomSheetScaffoldAnchorChangeHandler(
-    state: SheetState,
-    animateTo: (target: SheetValue, velocity: Float) -> Unit,
-    snapTo: (target: SheetValue) -> Unit,
-) = AnchorChangeHandler<SheetValue> { previousTarget, previousAnchors, newAnchors ->
-    val previousTargetOffset = previousAnchors[previousTarget]
-    val newTarget = when (previousTarget) {
-        SheetValue.Minimized -> SheetValue.PartiallyExpanded
-        SheetValue.Hidden, SheetValue.PartiallyExpanded -> SheetValue.PartiallyExpanded
-        SheetValue.Expanded -> if (newAnchors.containsKey(SheetValue.Expanded)) SheetValue.Expanded else SheetValue.PartiallyExpanded
-    }
-    val newTargetOffset = newAnchors.getValue(newTarget)
-    if (newTargetOffset != previousTargetOffset) {
-        if (state.swipeableState.isAnimationRunning) {
-            // Re-target the animation to the new offset if it changed
-            animateTo(newTarget, state.swipeableState.lastVelocity)
-        } else {
-            // Snap to the new offset value of the target if no animation was running
-            snapTo(newTarget)
-        }
-    }
-}
-
-@Composable
-fun StandardBottomSheetLayout(
-    modifier: Modifier = Modifier,
-    sheetOffset: () -> Float,
-    bottomSheet: @Composable (layoutHeight: Int) -> Unit,
-) {
-    SubcomposeLayout(modifier = modifier) { constraints ->
-        val layoutWidth = constraints.maxWidth
-        val layoutHeight = constraints.maxHeight
-
-        val sheetPlaceable = subcompose(0) {
-            bottomSheet(layoutHeight)
-        }[0].measure(constraints)
-        val sheetOffsetY = sheetOffset().roundToInt()
-        val sheetOffsetX = Integer.max(0, (layoutWidth - sheetPlaceable.width) / 2)
-
-        layout(layoutWidth, layoutHeight) {
-            sheetPlaceable.placeRelative(sheetOffsetX, sheetOffsetY)
-        }
-    }
-}
-
-private enum class BottomSheetScaffoldLayoutSlot { TopBar, Body, Sheet, Snackbar }
-
-/**
- * Defines the height values as fractions for the PartiallyExpanded and Expanded Sheet Values
- * of the bottom sheet
- *
- * @param partialHeightFraction fractional for the PartiallyExpanded height between [0,1]
- * @param fullHeightFraction fractional for the Expanded height between [0,1]
- */
-class SheetExpansionHeight(
-    val partialHeightFraction: Float = 0.5f,
-    val fullHeightFraction: Float = 1f
-)
