--- feature-forms/src/main/java/com/esri/arcgismaps/sample/featureforms/components/bottomsheet/Strings.kt
@@ -1,64 +1,0 @@
-/*
- * Copyright 2021 The Android Open Source Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- *
- * Modifications copyright (C) 2023 Esri Inc
- */
-
-package com.esri.arcgismaps.sample.featureforms.components.bottomsheet
-
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.Immutable
-import androidx.compose.ui.R
-import androidx.compose.ui.platform.LocalConfiguration
-import androidx.compose.ui.platform.LocalContext
-
-@Immutable
-@JvmInline
-value class Strings private constructor(
-    @Suppress("unused") private val value: Int = nextId()
-) {
-    companion object {
-        private var id = 0
-        private fun nextId() = id++
-
-        val NavigationMenu = Strings()
-        val CloseDrawer = Strings()
-        val CloseSheet = Strings()
-        val DefaultErrorMessage = Strings()
-        val ExposedDropdownMenu = Strings()
-        val SliderRangeStart = Strings()
-        val SliderRangeEnd = Strings()
-        val BottomSheetPartialExpandDescription = Strings()
-        val BottomSheetDismissDescription = Strings()
-        val BottomSheetExpandDescription = Strings()
-    }
-}
-
-@Composable
-fun getString(string: Strings): String {
-    LocalConfiguration.current
-    val resources = LocalContext.current.resources
-    return when (string) {
-        Strings.NavigationMenu -> resources.getString(R.string.navigation_menu)
-        Strings.CloseDrawer -> resources.getString(R.string.close_drawer)
-        Strings.CloseSheet -> resources.getString(R.string.close_sheet)
-        Strings.DefaultErrorMessage -> resources.getString(R.string.default_error_message)
-        Strings.ExposedDropdownMenu -> resources.getString(R.string.dropdown_menu)
-        Strings.SliderRangeStart -> resources.getString(R.string.range_start)
-        Strings.SliderRangeEnd -> resources.getString(R.string.range_end)
-        else -> ""
-    }
-}
